Hibernate  ORM   Test on  TestDB

+++ src/main/java/internetshop/niva/il/domain/Customer.java
@@ -1,19 +1,18 @@
 package internetshop.niva.il.domain;
 
-import javax.annotation.Generated;
 import javax.persistence.*;
 
 /**
  * Created by Igor on 2015.10.19..
  */
 @Entity
-@Table(name="customers")
+@Table(name = "customers")
 public class Customer {
 
     @Id
     @Column(name="CustomerID", columnDefinition = "INT(10)")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int userId;
+    private long userId;
 
     @Column(name = "FirstName", columnDefinition = "CHAR(40)")
     private String firstName;
@@ -26,7 +25,6 @@ public class Customer {
 
     @Column(name = "Email", columnDefinition = "CHAR(40)")
     private String email;
-
 
     public long getUserId() {
         return userId;
